+++ Sub-Terra/include/HumanPlayerController.h
@@ -13,13 +13,13 @@ private:
 	Point2 orientVel;
 	Point2 orientRot;
 	long double accum = 0.0;
-	float velocity = 10.0f;
+	Decimal velocity = 10.0;
 	bool moveForward = false, moveBackward = false, moveLeft = false, moveRight = false;
 protected:
 	virtual void Init() override;
 	virtual void Update(DeltaTicks &) override;
 public:
-	float time = 0.0f;
+	Decimal time = 0.0;
 
 	static bool IsSupported() { return true; }
 	HumanPlayerController(Polar *engine, const IDType object) : System(engine), object(object) {}
